+++ test/structural_typing/guts/f_explanations.clj
@@ -12,8 +12,4 @@
     ((:explainer oopsie) oopsie) => "path: [:a :b], leaf value: 33"))
 
 
-(future-fact "Hand-test range output - descend into 5, what incomplete range looks like")
-(future-fact "finish ALL output so that it chokes on non-sequential collection.")
-
-
 (future-fact "Add tests for all new explanations")